+++ api/http/options.go
@@ -15,6 +15,8 @@ import (
 type Options struct {
 	// HTTP client. Default is http.DefaultClient.
 	httpClient *http.Client
+	// doer is an http Doer - if set it overrides httpClient
+	doer Doer
 	// Flag whether http client was created internally
 	ownClient bool
 	// TLS configuration for secure connection. Default nil
@@ -27,7 +29,11 @@ type Options struct {
 // for HTTP requests. It will return the one that has been set using
 // SetHTTPClient or it will construct a default client using the
 // other configured options.
+// HTTPClient panics if SetHTTPDoer was called.
 func (o *Options) HTTPClient() *http.Client {
+	if o.doer != nil {
+		panic("HTTPClient called after SetHTTPDoer")
+	}
 	if o.httpClient == nil {
 		o.httpClient = &http.Client{
 			Timeout: time.Second * time.Duration(o.HTTPRequestTimeout()),
@@ -63,6 +69,32 @@ func (o *Options) SetHTTPClient(c *http.Client) *Options {
 // OwnHTTPClient returns true of HTTP client was created internally. False if it was set externally.
 func (o *Options) OwnHTTPClient() bool {
 	return o.ownClient
+}
+
+
+type Doer interface {
+	Do(*http.Request) (*http.Response, error)
+}
+
+// SetHTTPDoer will configure the http.Client that is used
+// for HTTP requests. If set to nil, this has no effect.
+//
+// Setting the HTTPDoer will cause the other HTTP options
+// to be ignored.
+func (o *Options) SetHTTPDoer(d Doer) *Options {
+	if d != nil {
+		o.doer = d
+		o.ownClient = false
+	}
+	return o
+}
+
+// HTTPDoer returns actual Doer if set, or http.Client
+func (o *Options) HTTPDoer() Doer {
+	if o.doer != nil {
+		return o.doer
+	}
+	return o.HTTPClient()
 }
 
 // TLSConfig returns tls.Config