+++ api/http/service.go
@@ -55,7 +55,7 @@ type service struct {
 	serverAPIURL  string
 	serverURL     string
 	authorization string
-	client        *http.Client
+	client        Doer
 }
 
 // NewService creates instance of http Service with given parameters
@@ -72,7 +72,7 @@ func NewService(serverURL, authorization string, httpOptions *Options) Service {
 		serverAPIURL:  serverAPIURL,
 		serverURL:     serverURL,
 		authorization: authorization,
-		client:        httpOptions.HTTPClient(),
+		client:        httpOptions.HTTPDoer(),
 	}
 }
 
@@ -127,7 +127,9 @@ func (s *service) DoHTTPRequestWithResponse(req *http.Request, requestCallback R
 	if len(s.authorization) > 0 {
 		req.Header.Set("Authorization", s.authorization)
 	}
-	req.Header.Set("User-Agent", http2.UserAgent)
+	if req.Header.Get("User-Agent") == "" {
+		req.Header.Set("User-Agent", http2.UserAgent)
+	}
 	if requestCallback != nil {
 		requestCallback(req)
 	}