+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+### Features
+[#256](https://github.com/influxdata/influxdb-client-go/pull/256) struct http.Service can be created with a 'Doer' interface
+
 ## 2.3.0 [2021-04-30]
 ### Breaking change
 [#253](https://github.com/influxdata/influxdb-client-go/pull/253) Interface 'Logger' extended with 'LogLevel() uint' getter.