+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Features
+[#256](https://github.com/influxdata/influxdb-client-go/pull/256) Allowing 'Doer' interface for HTTP requests
+
 ## 2.3.0 [2021-04-30]
 ### Breaking change
 [#253](https://github.com/influxdata/influxdb-client-go/pull/253) Interface 'Logger' extended with 'LogLevel() uint' getter.